+++ server/src/main/java/com/sprints/planning/dto/request/TaskCreateRequest.java
@@ -1,5 +1,6 @@
 package com.sprints.planning.dto.request;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.List;
 import java.util.Map;
@@ -12,8 +13,8 @@ public record TaskCreateRequest(
     String customer,
     String stream,
     List<String> participantIds,
-    Map<String, Integer> loads,
-    Map<String, Map<String, Integer>> allocations,
+    Map<String, BigDecimal> loads,
+    Map<String, Map<String, BigDecimal>> allocations,
     Map<String, String> notes,
     LocalDate releaseDate,
     String releaseSprintId,
